+++ runnable-hub/python/runnable/hub.py
@@ -1,39 +1,10 @@
 
-import dis
-from tracemalloc import start
-from pydantic import BaseModel
-from typing import List, Dict, Optional, Generic, TypeVar
+
 from datetime import datetime
-from enum import Enum
 import asyncio
 import uuid
 from abc import ABC, abstractmethod
-
-T = TypeVar('T')
-
-class RunnableStatus(Enum):
-    PENDING = "PENDING"
-    RUNNING = "RUNNING"
-    SUCCESS = "SUCCESS"
-    FAIL = "FAIL"
-    ERROR = "ERROR"
-
-class RunnableRequest(BaseModel):
-    runnableCode: str
-
-class RunnableResponse(BaseModel):
-    runnableCode: str
-
-class RunnableContext(BaseModel, Generic[T]):
-    request: T
-    response: Optional[RunnableResponse] = None
-    promiseRunnable: Dict[str, object] = {}
-    promiseResult: Dict[str, object] = {}
-    executeId: str
-    startTime: datetime
-    endTime: Optional[datetime] = None
-    status: RunnableStatus
-    errorMessage: Optional[str] = None
+from .context import RunnableRequest, RunnableContext, RunnableStatus
 
 class RunnableWorker(ABC):
     runnableCode = None
@@ -69,10 +40,12 @@ class RunnableWorkerDispatch():
         self.queue = queue
         self.dispatcher = dispatcher
 
+
 class RunnableHub():
 
-    def __init__(self):
+    def __init__(self, store):
         self.workers = {}
+        self.store = store
 
     def registerWorker(self, worker: RunnableWorker):
         self.workers[worker.runnableCode] = RunnableWorkerDispatch(
