+++ runnable-hub/python/runnable/store.py
@@ -1,0 +1,13 @@
+
+
+
+
+class RunnableLocalStore():
+    def __init__(self, path):
+        self.path = path
+
+    def save(self, fileName, content):
+        pass
+
+    def read(self, fileName):
+        pass
